move filter files around
