+++ vortex-compute/src/filter/mask.rs
@@ -8,16 +8,22 @@ use crate::filter::Filter;
 impl Filter for &Mask {
     type Output = Mask;
 
-    fn filter(self, mask: &Mask) -> Mask {
-        assert_eq!(self.len(), mask.len());
+    fn filter(self, selection_mask: &Mask) -> Mask {
+        assert_eq!(
+            selection_mask.len(),
+            self.len(),
+            "Selection mask length must equal the mask length"
+        );
 
-        match (self, mask) {
-            (Mask::AllTrue(_), _) => Mask::AllTrue(mask.true_count()),
-            (Mask::AllFalse(_), _) => Mask::AllFalse(mask.true_count()),
+        match (self, selection_mask) {
+            (Mask::AllTrue(_), _) => Mask::AllTrue(selection_mask.true_count()),
+            (Mask::AllFalse(_), _) => Mask::AllFalse(selection_mask.true_count()),
 
             (Mask::Values(_), Mask::AllTrue(_)) => self.clone(),
             (Mask::Values(_), Mask::AllFalse(_)) => Mask::new_true(0),
-            (Mask::Values(v1), Mask::Values(_)) => Mask::from(v1.bit_buffer().filter(mask)),
+            (Mask::Values(v1), Mask::Values(_)) => {
+                Mask::from(v1.bit_buffer().filter(selection_mask))
+            }
         }
     }
 }
